+++ Teensy/Blink/src/posxy_back.cpp
@@ -29,7 +29,7 @@ extern pid roll_control, pitch_control, yaw_control;
 IntervalTimer readSensors;
 IntervalTimer milli;
 
-volatile double roll, pitch, yaw, ax, ay, az;
+volatile float roll, pitch, yaw, ax, ay, az;
 float x, y, z;
 volatile unsigned long long time = 0;
 bool led_state;
@@ -135,11 +135,11 @@ int dig = 0;
 
 
 
-double  H,R,P,Y, H_ref, X_C, Y_C, R_MAX = pi/22.0 , P_MAX = pi/22.0;
-double M1,M2,M3,M4;
+float  H,R,P,Y, H_ref, X_C, Y_C, R_MAX = pi/22.0 , P_MAX = pi/22.0;
+float M1,M2,M3,M4;
 uint8_t haux = 0;
-double roll_off = 0 , pitch_off = 0, yaw_off = 0, x_off = 0, y_off = 0, z_off = 0;
-double roll_ref, pitch_ref, yaw_ref, x_ref, y_ref, z_ref;
+float roll_off = 0 , pitch_off = 0, yaw_off = 0, x_off = 0, y_off = 0, z_off = 0;
+float roll_ref, pitch_ref, yaw_ref, x_ref, y_ref, z_ref;
 long long pm = 0;
 
 
@@ -168,7 +168,7 @@ int _main(void){
 
         H /= cos(roll)*cos(pitch);
         
-        double rel = roll_ref/(pitch_ref + 0.0000001);
+        float rel = roll_ref/(pitch_ref + 0.0000001);
         
         if( fabs(rel) < 1  ){
